Move ZendeskManager and ZendeskSettings to dedicated package named zendesk
