+++ WooCommerce/src/main/kotlin/com/woocommerce/android/support/help/HelpViewModel.kt
@@ -2,7 +2,7 @@ package com.woocommerce.android.support.help
 
 import androidx.lifecycle.SavedStateHandle
 import com.woocommerce.android.extensions.formatResult
-import com.woocommerce.android.support.TicketType
+import com.woocommerce.android.support.zendesk.TicketType
 import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.util.WooLog
 import com.woocommerce.android.util.WooLogWrapper
